@@ -1,24 +1,29 @@
 package com.in28minutes.rest.basic.auth;
 
+import static org.springframework.security.config.Customizer.withDefaults;
+
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 @EnableWebSecurity
-public class SpringSecurityConfigurationBasicAuth extends WebSecurityConfigurerAdapter{
-	
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		http
-		.csrf().disable()	
-		.authorizeRequests()
-		.antMatchers(HttpMethod.OPTIONS,"/**").permitAll()
-				.anyRequest().authenticated()
-				.and()
-			//.formLogin().and()
-			.httpBasic();
+@EnableMethodSecurity
+public class SpringSecurityConfigurationBasicAuth {
+
+	@Bean
+	SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		http.authorizeHttpRequests(auth -> auth.requestMatchers(HttpMethod.OPTIONS, "/**").permitAll());
+
+		http.csrf().disable();
+
+		http.httpBasic(withDefaults());
+
+		return http.build();
 	}
+
 }
