@@ -1,76 +1,131 @@
 package com.in28minutes.rest.webservices.restfulwebservices.jwt;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.interfaces.RSAPrivateKey;
+import java.security.interfaces.RSAPublicKey;
+import java.util.UUID;
+
+import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.authentication.ProviderManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.oauth2.server.resource.OAuth2ResourceServerConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.jwt.JwtEncoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.web.SecurityFilterChain;
+
+import com.nimbusds.jose.JOSEException;
+import com.nimbusds.jose.jwk.JWKSet;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jose.jwk.source.JWKSource;
+import com.nimbusds.jose.proc.SecurityContext;
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-public class JWTWebSecurityConfig extends WebSecurityConfigurerAdapter {
+@EnableMethodSecurity
+public class JWTWebSecurityConfig {
 
-	@Autowired
-	private JwtUnAuthorizedResponseAuthenticationEntryPoint jwtUnAuthorizedResponseAuthenticationEntryPoint;
-
-	@Autowired
-	private UserDetailsService jwtInMemoryUserDetailsService;
-
-	@Autowired
-	private JwtTokenAuthorizationOncePerRequestFilter jwtAuthenticationTokenFilter;
-
-	@Value("${jwt.get.token.uri}")
-	private String authenticationPath;
-
-	@Autowired
-	public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(jwtInMemoryUserDetailsService).passwordEncoder(passwordEncoderBean());
+	@Bean
+	public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+		// https://github.com/spring-projects/spring-security/issues/1231
+		// https://docs.spring.io/spring-boot/docs/current/reference/html/data.html#data.sql.h2-web-console.spring-security
+		return httpSecurity
+				.authorizeHttpRequests(auth -> auth
+						.requestMatchers("/authenticate", "/hello-world", "/hello-world-bean").permitAll()
+						.requestMatchers(PathRequest.toH2Console()).permitAll() // h2-console is a servlet and NOT recommended for a production
+						.requestMatchers(HttpMethod.OPTIONS,"/**")
+						.permitAll()
+						.anyRequest()
+						.authenticated())
+				.csrf(AbstractHttpConfigurer::disable)
+				.sessionManagement(session -> session.
+						sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+				.oauth2ResourceServer(
+						OAuth2ResourceServerConfigurer::jwt)
+				.httpBasic(
+						Customizer.withDefaults())
+				.headers(header -> {
+					header.frameOptions().sameOrigin();
+				})
+				.build();
 	}
 
 	@Bean
-	public PasswordEncoder passwordEncoderBean() {
-		return new BCryptPasswordEncoder();
+	public AuthenticationManager authenticationManager(UserDetailsService userDetailsService) {
+		var authenticationProvider = new DaoAuthenticationProvider();
+		authenticationProvider.setUserDetailsService(userDetailsService);
+
+		return new ProviderManager(authenticationProvider);
 	}
 
 	@Bean
-	@Override
-	public AuthenticationManager authenticationManagerBean() throws Exception {
-		return super.authenticationManagerBean();
+	public UserDetailsService userDetailsService() {
+		UserDetails user = User.withUsername("in28minutes")
+				.password("{noop}dummy")
+				.authorities("read")
+				.roles("USER")
+				.build();
+
+		return new InMemoryUserDetailsManager(user);
 	}
 
-	@Override
-	protected void configure(HttpSecurity httpSecurity) throws Exception {
-		httpSecurity.csrf().disable().exceptionHandling()
-				.authenticationEntryPoint(jwtUnAuthorizedResponseAuthenticationEntryPoint).and().sessionManagement()
-				.sessionCreationPolicy(SessionCreationPolicy.STATELESS).and().authorizeRequests().anyRequest()
-				.authenticated();
+	@Bean
+	public JWKSource<SecurityContext> jwkSource() {
+		JWKSet jwkSet = new JWKSet(rsaKey());
 
-		httpSecurity.addFilterBefore(jwtAuthenticationTokenFilter, UsernamePasswordAuthenticationFilter.class);
-
-		httpSecurity.headers().frameOptions().sameOrigin() // H2 Console Needs this setting
-				.cacheControl(); // disable caching
+		return (((jwkSelector, securityContext)
+				-> jwkSelector.select(jwkSet)));
 	}
 
-	@Override
-	public void configure(WebSecurity webSecurity) throws Exception {
-		webSecurity.ignoring().antMatchers(HttpMethod.POST, authenticationPath)
-				.antMatchers(HttpMethod.OPTIONS, "/**")
-				.and().ignoring()
-				.antMatchers(HttpMethod.GET, "/" // Other Stuff You want to Ignore
-				).and().ignoring()
-				.antMatchers("/h2-console/**/**");// Should not be done in Production!
+	@Bean
+	JwtEncoder jwtEncoder(JWKSource<SecurityContext> jwkSource) {
+		return new NimbusJwtEncoder(jwkSource);
 	}
+
+	@Bean
+	JwtDecoder jwtDecoder() throws JOSEException {
+		return NimbusJwtDecoder
+				.withPublicKey(rsaKey().toRSAPublicKey())
+				.build();
+	}
+
+	@Bean
+	public RSAKey rsaKey() {
+		KeyPair keyPair = keyPair();
+
+		return new RSAKey
+				.Builder((RSAPublicKey) keyPair.getPublic())
+				.privateKey((RSAPrivateKey) keyPair.getPrivate())
+				.keyID(UUID.randomUUID().toString())
+				.build();
+	}
+
+	@Bean
+	public KeyPair keyPair() {
+		try {
+			var keyPairGenerator = KeyPairGenerator.getInstance("RSA");
+			keyPairGenerator.initialize(2048);
+			return keyPairGenerator.generateKeyPair();
+		} catch (Exception e) {
+			throw new IllegalStateException(
+					"Unable to generate an RSA Key Pair", e);
+		}
+	}
+
 }
